TERCERO/REDES2/PRACTICAS_V1/practica3/requirements.txt
appJar==0.94.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.1
pkg-resources==0.0.0
pycryptodome==3.9.7
requests==2.23.0
urllib3==1.25.8
